Initial implementation with XML parsing.
Properly generating a table of nodes for a plan, with all stuff filled in.
Includes the table spec generated by the `plan()` function. Probably some
typos and stuff; will have to add a bunch of tests to make sure they're right.

Also need to add some way to give each node a unique ID and to represent the
parent/child relationships. I'm not sure how I want to do that yet.