Add node IDs.
They are an MD5 of the concatenation of the back end PID and
clock_timestamp(). That should be unique enough.

Reqiured mocking md5() in `test/sql/base.sql` so that it always returns the
same values. Will add another test to make sure it concatenates what we think
it should concatenate.