Add optional `columns` parameter.
Contains a list of the columns to calculate. If this parameter is specified,
only the columns it lists will be populated with data. All others will be
`NULL`.

I'm not thrilled with the implementation, but it is at least self-contained.
We can always refactor `parse_node()` later to be more efficient. The
performance when the `columns` parameter is not passed (or is `NULL`) is
mostly unaffected.